requirements.txt
Pillow==9.1.0
pydantic==1.9.0
requests==2.27.1
setuptools==62.0.0
